feat: issue warning when gno test package can't be resolved (#1845)
Pretty straightforward one line change. If `gno test` fails to find a
gno.mod field or package path in gno root, the result will be a panic
with the message `panic: unexpected unreal object`, which makes it
pretty difficult to figure out what must be done to correct the issue.
The warning gives a solution on how to fix this.