feat(stdlibs): Fuzz for String (#1809)
## Description

I have implemented basic Fuzz functions in the testing package and a
function to apply fuzzing to string data.

To transform string data, I used genetic algorithms. Additionally, I
implemented a Uniform distribution to generate random values and then
used the central limit theorem to approximate a normal distribution,
thereby enabling the creation of random booleans in the random file.

In this PR, only the string type is handled. I plan to support other
types in the future.