docs: add gnoclient reference & how-to guide (#1562)
## Description

This PR adds the reference documentation to the newly added
[gnoclient](https://github.com/gnolang/gno/pull/1047).
The reference docs for gnoclient are mostly generated by using a godoc
to markdown tool, with slight modifications.

It also adds a how-to guide on how to connect to a Gno.land chain from
Go.

Along the way: 
- Adds index pages & minor doc improvements for `tm2-js` & `gno-js`
- Updates godoc comments in client
- Changes `gnoclient` function signatures to all be on the same type of
receiver (all tests pass). I realize this could've been a different PR
but its a small change with no functional/practical impact.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>