fix: hardcode max vm cycles in keeper (#1807)
<!-- please provide a detailed description of the changes made in this
pull request. -->
This hardcodes the maximum VM cycles in the keeper to ten million, the
same that is currently being used from genesis. It doesn't seem like
there is a reason why we should want people to be able to adjust this.
<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [x] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>