feat(examples): add simple userbook realm (#1949)
<!-- please provide a detailed description of the changes made in this
pull request. -->

## Description

This PR arose out of a necessity to have a simple, permissionless realm
for demonstration purposes. The idea is that people can sign up once
with their address, and the Render function will paginate 50 addresses
per page, as well as include a UI to go to the next/previous page. This
realm will serve 3 purposes:
- Show how to do pagination
- Show how to use Render in harmony with `gnoweb`
- Exist as a simple entry point for new developers testing out the
chain. This realm will be used as an example in the Getting started
section in the docs.


<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: deelawn <dboltz03@gmail.com>