feat: Oracle packages and Github verification realm (#1568)
Relates to #1133 

Note: the unit tests are missing from the gnorkle packages. This will
not be merged without unit tests. They will be written once the design
has been reviewed so no time is wasted writing and rewriting them should
it be determined that structural changes need to occur.

Note 2: there are two readme files included in this PR so maybe read
those first.

The PR is meant to be a first attempt at implementing a gno package that
allows realms to integrate oracle functionality. There is a gh (github)
application bundled in this PR that is an oracle and will be able to
handle github handle -> gno address verification requests.

Everything at the gno package (/p) level falls within a v1 path because
I think this package still has a lot of room for improvements.

Some things that need to be done before marking this ready for review:
- [x] remove old first iteration code
- [x] convert newest prototyped code from go to gno and restructure
subdirectories for compatibility
- [x] add more comments and doc strings
- [x] create readmes
- [x] add unit tests
- [x] add gh verification realm tests
- [x] make a way for the owner or an orkle instance to manage both
instance and individual feeds' whitelists

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Morgan <git@howl.moe>