docs(getting-started): local setup with gnodev (#1936)
<!-- please provide a detailed description of the changes made in this
pull request. -->

## Description

This PR introduces a change to the Getting Started section in the docs.
As discussed with @moul, it makes sense to have `gnodev` take the
spotlight in the "Local setup" section.

With this PR, I've modified the "Local setup" section to reflect the use
of `gnodev`. I've moved the "Setting up a local chain" & "Premining
balances" out of this section, as they will belong better in a new
section which will contain new chain initialization flows.

~Dep. on https://github.com/gnolang/gno/pull/1949~

Latest preview:
https://www.loom.com/share/a9d354b7234843fb8108f1698cc28b5c?sid=92850b75-5153-4f8b-aa0d-bb9d3aa0d6fd

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>