style: logo and font update (#2132)
<!-- please provide a detailed description of the changes made in this
pull request. -->

**This PR aims to replace the GnoLand logo by the new one and to swap
monospaced body font by a non-mono one.**

- From what I understood, we are gonna use only one version of the logo
for both light and dark theme, so I had to update the header background
color for visibility and contrast matter. I already made some iterations
on the colors with Eunice. I also adapted the horizontal logo -
navigation balance by updating the margin.
- I replaced the standalone `logo.svg` by a SVG `embedded` version of
the new one. I know there is pros and cons about that but it's temporary
before the revamp of `gnoweb`, when we will be able to discuss it. As we
want to go fast, I think this is easier to customize as we can iterate
by tweaking few lines of css instead of re-uploading new images (and
hosting multiples related to different theme). It's also faster to
display and handy to interact with.
- I renamed the go template to make it more generic since we also use
the logo in the footer.

- Finally, I got rid of monospaced font (for body text only) to make a
test with `ui-system` one. It's the one used for titles right now (the
main difference with title being the font-size, font-weight and
letter-spacing). Using a `ui-system` font means all OS will have
different font displayed (eg. San Francisco for MacOS / Roboto for
Android / Ubuntu font for Linux...), what is not the best for branding
matter, but it is very fast to display (actually nothing to load, so no
`FOUT` issue), common to see for users (default os font) and maybe a
good choice as the branding is not fully finished yet. This would be a
consensual but good way to start because this font is already a part of
the website.
- We can also quickly make a test with `Inter` font, that is very
simple, elegant and used in many of our websites, and otherwise looking
for other fonts to tests.

## Screenshots:
### Light theme: 
![127 0 0
1_8888_](https://github.com/gnolang/gno/assets/7622257/9af2b1f6-bb2b-4514-90e1-0ca74e42024f)

### dark theme:
![127 0 0 1_8888_
(1)](https://github.com/gnolang/gno/assets/7622257/53372b03-bc67-444b-a54a-5b1fbb62824a)