Replace value with placeholder.
By having "ADDRESS" in the field, it leads users to issues when they paste in their address, as they assume only their address will be in the field. 

Currently it might end up with ADDRESS appended to their actual address, which causes confusing when the help generation does not work because of this.